+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Numbers in parentheses after entries refer to issues in the [GitHub issue tracke
 - `isqlite update` now treats `NULL` as a special value, in line with `isqlite iupdate`.
 - `isqlite delete` now displays the string values of foreign keys when printing a row prior to deletion.
 
+### Fixed
+- In `isqlite icreate`, empty values are ignored instead of being inserted as NULLs or empty strings. This is useful when the column definition specifies a default value.
+
 
 ## [1.4.0] - 2021-11-17
 ### Added